+++ src/ShellCheck/ASTLib.hs
@@ -776,6 +776,7 @@ prop_executableFromShebang8 = executableFromShebang "/usr/bin/env --split-string
 prop_executableFromShebang9 = executableFromShebang "/usr/bin/env foo=bar dash" == "dash"
 prop_executableFromShebang10 = executableFromShebang "/bin/busybox sh" == "busybox sh"
 prop_executableFromShebang11 = executableFromShebang "/bin/busybox ash" == "busybox ash"
+prop_executableFromShebang12 = executableFromShebang "/sbin/env irix-sh" == "irix-sh"
 
 -- Get the shell executable from a string like '/usr/bin/env bash'
 executableFromShebang :: String -> String