+++ src/ShellCheck/Data.hs
@@ -170,9 +170,15 @@ shellForExecutable name =
         "ksh88" -> return Ksh
         "ksh93" -> return Ksh
         "oksh"  -> return Ksh
+        "irix-sh" -> return IrixSh
         _ -> Nothing
 
 flagsForRead = "sreu:n:N:i:p:a:t:"
+
+-- IRIX sh uses -p to read from the coprocess, without an option argument.
+flagsForReadFor IrixSh = "sreu:n:N:i:pa:t:"
+flagsForReadFor _ = flagsForRead
+
 flagsForMapfile = "d:n:O:s:u:C:c:t"
 
 declaringCommands = ["local", "declare", "export", "readonly", "typeset", "let"]